Rename the singer change handler.
